cambiando double a float

+++ db/migrate/20180906210748_change_porcentaje_cotizacion_isapre_to_be_double.rb
@@ -1,6 +1,6 @@
 
 class ChangePorcentajeCotizacionIsapreToBeDouble < ActiveRecord::Migration[5.1]
   def change
-  	change_column :empleados, :porcentajeCotizacionIsapre, :double
+  	change_column :empleados, :porcentajeCotizacionIsapre, :float
   end
 end
